+++ dns/src/lib.rs
@@ -53,16 +53,13 @@
 //!
 //! * `eh0`: Passthrough to [`w5500-hl`].
 //! * `eh1`: Passthrough to [`w5500-hl`].
-//! * `ip_in_core`: Passthrough to [`w5500-hl`].
-//! * `std`: Passthrough to [`w5500-hl`].
 //! * `defmt`: Enable logging with `defmt`. Also a passthrough to [w5500-hl].
 //! * `log`: Enable logging with `log`.
 //!
 //! [`w5500-hl`]: https://crates.io/crates/w5500-hl
-//! [`std::net`]: https://doc.rust-lang.org/std/net/index.html
 //! [Wiznet W5500]: https://www.wiznet.io/product-item/w5500/
 #![cfg_attr(docsrs, feature(doc_cfg), feature(doc_auto_cfg))]
-#![cfg_attr(all(not(feature = "std"), not(test)), no_std)]
+#![cfg_attr(not(test), no_std)]
 #![deny(unsafe_code)]
 #![warn(missing_docs)]