+++ regsim/CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Removed
+- Removed the `ip_in_core` feature, `Ipv4Addr` and `SocketAddrV4` are now provided by `core::net`.
 - Removed the `async` feature, asynchronous implementations are enabled by default.
 
 ## [0.1.0-alpha.9] - 2023-08-06